+++ derive_tests/tests/length.rs
@@ -96,7 +96,7 @@ fn value_out_of_length_fails_validation() {
     let errs = err.field_errors();
     assert_eq!(errs.len(), 1);
     assert_eq!(errs[0].code(), "length");
-    assert_eq!(errs[0].params()["value"], "");
+    assert_eq!(errs[0].params()["actual"], "");
     assert_eq!(errs[0].params()["min"], 5);
     assert_eq!(errs[0].params()["max"], 10);
 }
@@ -135,8 +135,6 @@ fn can_specify_message_for_length() {
 
 #[test]
 fn can_validate_array_for_length() {
-    use serde_json::Value;
-
     #[derive(Debug, Validate)]
     struct TestStruct {
         #[validate(length(min = 5, max = 10))]
@@ -152,10 +150,6 @@ fn can_validate_array_for_length() {
     let errs = err.field_errors();
     assert_eq!(errs.len(), 1);
     assert_eq!(errs[0].code(), "length");
-    assert_eq!(
-        errs[0].params()["value"],
-        Value::Array(vec![Value::String(String::new())])
-    );
     assert_eq!(errs[0].params()["min"], 5);
     assert_eq!(errs[0].params()["max"], 10);
 }