+++ derive_tests/tests/phone.rs
@@ -50,7 +50,7 @@ fn can_specify_code_for_phone() {
     let errs = err.field_errors();
     assert_eq!(errs.len(), 1);
     assert_eq!(errs[0].code(), "oops");
-    assert_eq!(errs[0].params()["value"], "bob");
+    assert_eq!(errs[0].params()["actual"], "bob");
 }
 
 #[test]